backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jordan Robin",email = "jordanrobin100@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "django (>=6.0,<7.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-stubs (>=3.16.6,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
